Moving city hex polygon lookups to lake
The bucket on edge got deleted! This was a bit nasty because of code DRYness, I had to change any scripts using the get_layers function. Probably could have done something cleverer inside the function to assume EDGE unless otherwise specified, but it is written that 'explict is better than implicit'.